feat: add disable rule cli option
