+++ src/linter/rules/index.mjs
@@ -1,0 +1,14 @@
+'use strict';
+
+import { invalidChangeVersion } from './invalid-change-version.mjs';
+import { missingChangeVersion } from './missing-change-version.mjs';
+import { missingIntroducedIn } from './missing-introduced-in.mjs';
+
+/**
+ * @type {Record<string, import('../types').LintRule>}
+ */
+export default {
+  'invalid-change-version': invalidChangeVersion,
+  'missing-change-version': missingChangeVersion,
+  'missing-introduced-in': missingIntroducedIn,
+};
